working on photo capture
added class Utils with methods dispatchTakePictureIntent createImageFile convertToBase64
calling take photo onCLickListener

+++ app/src/main/java/com/example/cettorre/animalsshelter/application/dto/AnimalDTO.java
@@ -11,8 +11,8 @@ public class AnimalDTO {
     private boolean hasChip;
     private Date date;
     private String photoB64;
-    private float latitude;
-    private float longitude;
+    private double latitude;
+    private double longitude;
 
     public AnimalDTO(Animal animal) {
         this.name = animal.getName();
